// spike/api-ergonomics/src/types.ts
export interface RunInput {
  inputs?: Record<string, string | number | boolean>;
  env?: Record<string, string>;
  github?: Partial<GitHubContext>;
  runner?: Partial<RunnerContext>;
  secrets?: Record<string, string>;
  jobStatus?: 'success' | 'failure' | 'cancelled';
}

export interface GitHubContext {
  repository: string;
  repositoryOwner: string;
  sha: string;
  ref: string;
  refName: string;
  refType: string;
  actor: string;
  eventName: string;
  workflow: string;
  runId: string;
  runNumber: string;
  runAttempt: string;
  serverUrl: string;
  apiUrl: string;
  [key: string]: unknown;
}

export interface RunnerContext {
  os: string;
  arch: string;
  name: string;
  temp: string;
  toolCache: string;
}